// src/types/template.ts
export interface OceaniaServiceItem {
    id: string;
    title: string;
    description: string;
    icon?: string;
    image?: string;
}

export interface OceaniaServicesProps {
    badgeText: string;
    title: string;
    description: string;
    services: OceaniaServiceItem[];
}